Refactor and fix the cover block unwrapping of columns
